[Update] tasks/ImportData

+++ application/tasks/controller/ImportData.php
@@ -18,7 +18,8 @@ class ImportData
         try {
             Log::record('[ImportData] start');
 
-            $res = Db::table('ems_main_engine')->distinct(true)->field('category')->select();
+            $res = Db::table('ems_main_engine')->distinct(true)->field('category')
+                ->whereNotNull('category')->select();
 
             $saveArr = [];
             foreach ($res as $index => $item) {
@@ -26,22 +27,29 @@ class ImportData
                     $saveArr[] = $value;
                 }
             }
-            // 先删除ems_const表
-            try {
-                Db::table('ems_const')->delete(true);
-            } catch (Exception $e) {
-                Log::record('delete ems_const fail ... '. $e->getMessage());
-            }
-            Log::record('start update ems_const ...');
 
-            // 重置序列
-            Db::execute('ALTER TABLE ems_const AUTO_INCREMENT = 1;');
-            // 插入数据
-            foreach ($saveArr as $value) {
-                if (!empty($value)) {
-                    Db::table('ems_const')
-                        ->data(['id'=>null,'name'=>$value])
-                        ->insert();
+            // 获取ems_const表数量
+            $total = Db::table('ems_const')->count();
+            if ($total == count($saveArr)) {
+                Log::record('ems_const has no need to update ... total = '. $total);
+            } else {
+                // 先删除ems_const表
+                try {
+                    Db::table('ems_const')->delete(true);
+                } catch (Exception $e) {
+                    Log::record('delete ems_const fail ... '. $e->getMessage());
+                }
+                Log::record('start update ems_const ...');
+
+                // 重置序列
+                Db::execute('ALTER TABLE ems_const AUTO_INCREMENT = 1;');
+                // 插入数据
+                foreach ($saveArr as $value) {
+                    if (!empty($value)) {
+                        Db::table('ems_const')
+                            ->data(['id'=>null,'name'=>$value])
+                            ->insert();
+                    }
                 }
             }
         } catch (Exception $e) {
